requirements.txt
aiogram
lightdb
python-dotenv
octodiary
gitpython